requirements.txt
flask==2.0.1
flask-restx==0.5.1
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.46
python-dotenv==0.19.1
openai==1.3.5
pdfplumber==0.7.0
werkzeug==2.0.2